Change documentation and tests to new interface for stub(), verify() and inspect()
Method calls will now be called on the mock itself, rather than separating the method call from the mock. This makes for a cleaner syntax. But the call needs to be wrapped in a coderef so we can catch exception throws and to track whether the call was real or not.