Add prototype to stub(), change response methods
Responses are no longer called as methods but as functions (but the old methods are retained for backwards compatibility). The reason for this change is to trim down the syntax. There is no longer any need for the arrows (->). Furthermore, with the response functions the parentheses may be omitted for single responses where there is no ambiguity of argument groupings.