Fix #15 - Loof for TODO in the right place
In CalledOk '$Test::Builder::Level' is altered before calling ok(). This
is not itself a problem, the problem is that the alteration is not
restored before calling $TB->TODO, as such it is looking for TODO in the
wrong place.

The test to ensure the diagnostics don't show up was incorrect and
checked that they do show up. These 2 errors together resulted in a test
passing when it should nto have.

This was discovered with an upcoming Test::More dev release. With this
path the tests pass on both new and old versions.